requirements.txt
flask==3.1.1
flask-cors==4.0.2
python-dotenv==1.0.1
openai==1.51.0
requests==2.32.4
